scripts/requirements.txt
httpx>=0.27.0
beautifulsoup4==4.12.2
lxml==5.1.0
supabase>=2.10.0
python-dotenv==1.0.0
